Use binary for building docker image
In the previous workflow, we were building and testing twice: once for
the docker image and once for the test step. Now, we'll build a debug
version in the test step, then build a release version and finally copy
that into the docker image. This would heavily reduce our build times.

Signed-off-by: Sayan Nandan <nandansayan@outlook.com>