Use binary for building docker image (#40)
* Use binary for building docker image

In the previous workflow, we were building and testing twice: once for
the docker image and once for the test step. Now, we'll build a debug
version in the test step, then build a release version and finally copy
that into the docker image. This would heavily reduce our build times.

* Ignore specific files in target to speed up builds

* Build image only when pushed to next or tagged

* Fix build condition

Since both conditions have to evaluate to true, we'll need to use refs
while also matching against both `push` and `tag` events

Signed-off-by: Sayan Nandan <nandansayan@outlook.com>