Enable the table to be frozen
A custom conditional variable was added to enable writes on the table to
be frozen temporarily. To do this, the caller must call `unlock_writes`
and then wait to get a `TableLockStateGuard`. The caller can do anything
with the table at this point.

What we do need to do is investigate race conditions on the lock_writes
member function of `SkymapInner`.