Add `OccupiedEntry` struct for easier interfacing
This idea was taken from `std`'s HashMap having an entry API. It is
far easier to call .key() and .value() than checking it manually with
patterns. However, one thing to keep in mind is that iteration is, and
will be inefficient in Hashtables especially because keys can be very
far apart.