Unaligned read instead of non-overlapping copy
It is bad to assume that the `start_ptr` will be aligned
when copying to the zeroed array. Instead we should just
do an unaligned read  and then unsafely cast the array.

It is not just bad: the start_ptr will never be aligned.

Deref instead of attempting a non-overlapping copy