Fix state machines, refactor schema and add `ASTNode` def and impls
This is one huge commit that fixes a bunch of issues we've had with
the schema and surrounding impls. Here are some of them
- The type metadata syn parse was incredibly broken. More than
one trailing comma or other punctuation following a trailing symbol
(as in a comma) is illegal
- The `*_full` set of methods were repetitive and annoying. That has
been fixed (not the most elegant solution, but it works and fits
well)
- As a result of the first issue, fuzz targets kept blowing up,
rightfully. This has been fixed too. The fuzz needs to be more
deterministic covering all permutations, and this is something I'll
fix up ahead.

Finally, I can now happily say that the QL impl is rock solid. There is
another possible bug that I'm tracking down, but that will be fixed.