Remove the shared impl of seq index
I actually realized this is a terrible abstraction the day I implemented
it and have since been planning to remove it.

The issue is with aliasing and potential corruption in a multi-threaded
environment. Even though we have never used it in such a context (aka
triggering UB) but we can potentially make an *accidental* use of it;
don't take it lightly: it's like modifying data while someone else is
reading it. That's not bad, that's a nightmare.