Rewrite journal: recovery impl and driver init
On initialization the driver will no longer attempt to truncate the
close event but instead use a truly AO style and utilize a new
directive.

We also now have a basic externally triggered on write and automatically
trigger on read recovery system in place. However, there are several
scopes of improvement to the implementation, such as tolerating even
more inconsistencies in the data.