Use driver contexts and improve model data batch recovery
This implementation revises the way we handle errors for both data
delta sync and that of direct driver events.

For the GNS: we now check the instance state before doing anything
at all. This is much better because we don't want to aggravate the
on-disk situation which is probably already in a very bad layout.

The same for the model driver: we check the instance's state before
doing anything at all. We also utilized a driver state to check
what was synced, following the changes that were made in an earlier
commit (skytable/skytable@e28d94e) where we only load the delta
that failed to sync back to the model's delta state.